Whatsapp/requirements.txt
flask
twilio
python-dotenv
geopandas
shapely
